Refactor into own class
