+++ photon-core/src/main/java/org/photonvision/common/logging/Logger.java
@@ -17,9 +17,7 @@
 
 package org.photonvision.common.logging;
 
-import edu.wpi.first.util.RuntimeDetector;
 import java.io.*;
-import java.nio.file.Files;
 import java.nio.file.Path;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
@@ -31,7 +29,6 @@ import org.photonvision.common.configuration.PathManager;
 import org.photonvision.common.dataflow.DataChangeService;
 import org.photonvision.common.dataflow.events.OutgoingUIEvent;
 import org.photonvision.common.util.TimedTaskManager;
-import org.photonvision.jni.QueuedFileLogger;
 
 /** TODO: get rid of static {} blocks and refactor to singleton pattern */
 public class Logger {
@@ -40,8 +37,8 @@ public class Logger {
 
     private static final UILogAppender uiLogAppender = new UILogAppender();
 
-    // TODO why's the logger care about this? split it out
-    private static KernelLogListener klogListener = null;
+    // // TODO why's the logger care about this? split it out
+    // private static KernelLogLogger klogListener = null;
 
     static {
         levelMap.put(LogGroup.Camera, LogLevel.INFO);
@@ -59,10 +56,6 @@ public class Logger {
         addFileAppender(PathManager.getInstance().getLogPath());
 
         cleanLogs(PathManager.getInstance().getLogsDir());
-    }
-
-    public static void addKlongListener() {
-        klogListener = new KernelLogListener();
     }
 
     public static final String ANSI_RESET = "\u001B[0m";
@@ -295,38 +288,6 @@ public class Logger {
             return sw.toString();
         } catch (IOException ioe) {
             throw new IllegalStateException(ioe);
-        }
-    }
-
-    private static class KernelLogListener {
-        QueuedFileLogger listener = null;
-        Logger logger = new Logger(KernelLogListener.class, LogGroup.General);
-
-        public KernelLogListener() {
-            if (RuntimeDetector.isLinux()) {
-                logger.info("Listening for klogs on /var/log/dmesg ! Boot logs:");
-
-                try {
-                    var bootlog = Files.readAllLines(Path.of("/var/log/dmesg"));
-                    for (var line : bootlog) {
-                        logger.log(line, LogLevel.DEBUG);
-                    }
-                } catch (IOException e) {
-                    logger.error("Couldn't read /var/log/dmesg - not printing boot logs");
-                }
-
-                listener = new QueuedFileLogger("/var/log/kern.log");
-            } else {
-                System.out.println("NOT for klogs");
-            }
-
-            TimedTaskManager.getInstance().addTask("outputPrintk", this::outputNewPrintks, 1000);
-        }
-
-        public void outputNewPrintks() {
-            for (var msg : listener.getNewlines()) {
-                logger.log(msg, LogLevel.DEBUG);
-            }
         }
     }
