+++ photon-server/src/main/java/org/photonvision/Main.java
@@ -33,6 +33,7 @@ import org.photonvision.common.dataflow.networktables.NetworkTablesManager;
 import org.photonvision.common.hardware.HardwareManager;
 import org.photonvision.common.hardware.PiVersion;
 import org.photonvision.common.hardware.Platform;
+import org.photonvision.common.logging.KernelLogLogger;
 import org.photonvision.common.logging.LogGroup;
 import org.photonvision.common.logging.LogLevel;
 import org.photonvision.common.logging.Logger;
@@ -437,8 +438,8 @@ public class Main {
         Logger.setLevel(LogGroup.General, logLevel);
         logger.info("Logging initialized in debug mode.");
 
-        // after native libraries are loaded :(
-        Logger.addKlongListener();
+        // Add Linux kernel log->Photon logger
+        KernelLogLogger.getInstance();
 
         // Add CSCore->Photon logger
         PvCSCoreLogger.getInstance();
